refactor(jira-channels): Moved package and added copyright header.
