--- channel/src/main/java/com/synopsys/integration/alert/channel/jira2/server/JiraServerIssueCommenter.java
@@ -1,20 +1,0 @@
-package com.synopsys.integration.alert.channel.jira2.server;
-
-import com.synopsys.integration.alert.channel.api.issue.send.IssueTrackerIssueResponseCreator;
-import com.synopsys.integration.alert.channel.jira2.common.JiraIssueCommenter;
-import com.synopsys.integration.alert.common.persistence.model.job.details.JiraServerJobDetailsModel;
-import com.synopsys.integration.jira.common.cloud.service.IssueService;
-
-public class JiraServerIssueCommenter extends JiraIssueCommenter {
-    private final JiraServerJobDetailsModel distributionDetails;
-
-    public JiraServerIssueCommenter(IssueTrackerIssueResponseCreator<String> issueResponseCreator, IssueService issueService, JiraServerJobDetailsModel distributionDetails) {
-        super(issueResponseCreator, issueService);
-        this.distributionDetails = distributionDetails;
-    }
-
-    @Override
-    protected boolean isCommentingEnabled() {
-        return distributionDetails.isAddComments();
-    }
-}
